(maint) Ensure handle_puppet_run_returned_exit_code can use a Range
Previously if none of the catch_* and expect_* options where set the
get_acceptable_puppet_run_exit_codes method would emit a Range (0..256).  However
this raised errors in the handle_puppet_run_returned_exit_code method because
you cannot do a `.join` on a Range.  This commit changes the output if it as an
array;  Joining if it is an array and converting to a string for other types.

Tests have been added for this behavior.