(QA-3461) - Making `execute_manifest_on` compatible with `handle_puppet_run_returned_exit_code`
Currently functionality does not exist to handle Arrays of returned exit
codes. This will happen in a case where there are more than one agent
responding.

This commit checks and if the returned `res` is an Array it is able to
loop through each exit code to check if they are valid or not.